fix: Correct storedDisplay for subtract and divide
This fixes an issue where the storedDisplay was showing calculations
improperly for minus and divide. Instead of using the previous operand +
operator, It now shows the full previous calculation.
ex: shows '65 - 5' instead of '5 -' after calculating 65 - 5.

It also no longer shows the same value as the currentNumber because it
now updates before equals calculates the new currentNumber.